refactor: standardize icon imports and file naming conventions across components for improved consistency
