+++ sheet/components/icons/delete.tsx
